synclient: fix size print

+++ src/client/SyntheticClient.cc
@@ -1435,7 +1435,7 @@ int SyntheticClient::full_walk(string& basedir)
       // print
       char *tm = ctime(&st.st_mtime);
       tm[strlen(tm)-1] = 0;
-      printf("%llx %c%c%c%c%c%c%c%c%c%c %2d %5d %5d %8d %12s %s\n",
+      printf("%llx %c%c%c%c%c%c%c%c%c%c %2d %5d %5d %8llu %12s %s\n",
 	     (long long)st.st_ino,
 	     S_ISDIR(st.st_mode) ? 'd':'-',
 	     (st.st_mode & 0400) ? 'r':'-',
@@ -1449,7 +1449,7 @@ int SyntheticClient::full_walk(string& basedir)
 	     (st.st_mode & 01) ? 'x':'-',
 	     (int)st.st_nlink,
 	     (int)st.st_uid, (int)st.st_gid,
-	     (int)st.st_size,
+	     (long long unsigned)st.st_size,
 	     tm,
 	     file.c_str());
